Swoff 321 2 (#1801)
* several apps impl

* adjusting panels

* adjusting sidemenu

* fixing queries, and schema references

* fixing audit thread sync issue

* updating audit package

* fixing multiple download attachments and quick sync across panels

* adding reset capabilities

* fixing titles, commenting out, moving reset sync to long press

+++ softwrench.sW4.WebMobile/www/Content/Mobile/scripts/services/login/securityservice.js
@@ -159,20 +159,29 @@
          * 
          * @return Promise resolved with the logged out user 
          */
-        const logout = function (clearSettings =false) {
+        const logout = function (clearSettings =false, clearCookies = true) {
             // invalidate current session
-            const current = localStorageService.remove(config.authkey); 
+
+            let current = null;
+
+            if (clearCookies) {
+                current = localStorageService.remove(config.authkey);
+            } else {
+                current = localStorageService.get(config.authkey);
+            }
+            
             // making sure the previous user is always the last "active" user
             if (!!current) {
                 localStorageService.put(config.previouskey, current);
             }
             $rootScope.$broadcast($event("logout"), current);
             const toAvoid = clearSettings ? [] : ["Settings"];
+            const promises = [dao.resetDataBase(toAvoid)];
+            if (clearCookies) {
+                promises.push(cookieService.clearCookies());
+            }
 
-            return $q.all([
-                dao.resetDataBase(toAvoid),
-                cookieService.clearCookies() // clear cookies 
-            ]).then(() => {
+            return $q.all(promises).then(() => {
                 $ionicHistory.clearCache(); // clean cache otherwise some views may remain after a consecutive login
                 cleanLocalStorage(); // clean non-blacklisted localstorage entries used by apps as cache
                 indexCreatorService.dropIndexes(); //drop sql indexes to allow fast sync insertion
